pyproject.toml
[tool.poetry]
name = "cyclarity-in-vehicle-sdk"
version = "0.1.0"
description = ""
authors = ["Gal Rahamim <gal.rahamim@cymotive.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
cyclarity-sdk = "1.0.49"
nanobind = "^2.1.0"
typing-extensions = "^4.12.2"
adb-shell = "^0.4.4"
serial = "^0.0.97"
paramiko = "^3.4.1"
doipclient = "^1.1.1"
can-isotp = "^2.0.6"
udsoncan = "^1.23.1"
py_pcapplusplus = {path = "./py_pcapplusplus"}
cmake = "^3.30.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
